Move tests to 'unit_tests' to prepare ML Testing folder
