+++ dwi_ml/unit_tests/test_batch_loader.py
@@ -8,8 +8,8 @@ from torch.utils.data.dataloader import DataLoader
 
 from dwi_ml.data.dataset.multi_subject_containers import MultiSubjectDataset
 from dwi_ml.models.main_models import MainModelOneInput
-from dwi_ml.tests.utils.expected_values import TEST_EXPECTED_NB_STREAMLINES
-from dwi_ml.tests.utils.data_and_models_for_tests import (
+from dwi_ml.unit_tests.utils.expected_values import TEST_EXPECTED_NB_STREAMLINES
+from dwi_ml.unit_tests.utils.data_and_models_for_tests import (
     create_test_batch_sampler, create_batch_loader, fetch_testing_data)
 
 SPLIT_RATIO = 0.5
