+++ dwi_ml/unit_tests/test_eos_sos_and_class.py
