+++ dwi_ml/unit_tests/visual_tests/analyze_batch_loader_visually.py
@@ -13,7 +13,7 @@ from dipy.io.streamline import save_tractogram
 
 from dwi_ml.data.dataset.multi_subject_containers import MultiSubjectDataset
 from dwi_ml.models.main_models import MainModelOneInput
-from dwi_ml.tests.utils.data_and_models_for_tests import (
+from dwi_ml.unit_tests.utils.data_and_models_for_tests import (
     create_test_batch_sampler, create_batch_loader, fetch_testing_data,
     ModelForTest)
