+++ scripts_python/l2t_visualize_error.py
@@ -7,7 +7,7 @@ from scilpy.io.utils import assert_inputs_exist, assert_outputs_exist
 from torch.nn.utils.rnn import unpack_sequence
 
 from dwi_ml.models.projects.learn2track_model import Learn2TrackModel
-from dwi_ml.tests.visual_tests.visualise_error_sft import \
+from dwi_ml.unit_tests.visual_tests.visualise_error_sft import \
     build_argparser_visu_error, \
     prepare_batch_visu_error, save_output_with_ref
