+++ scripts_python/tests/test_all_steps_learn2track.py
@@ -6,10 +6,10 @@ import pytest
 
 import torch
 
-from dwi_ml.tests.utils.expected_values import (
+from dwi_ml.unit_tests.utils.expected_values import (
     TEST_EXPECTED_VOLUME_GROUPS,
     TEST_EXPECTED_STREAMLINE_GROUPS, TEST_EXPECTED_SUBJ_NAMES)
-from dwi_ml.tests.utils.data_and_models_for_tests import fetch_testing_data
+from dwi_ml.unit_tests.utils.data_and_models_for_tests import fetch_testing_data
 
 data_dir = fetch_testing_data()
 
@@ -37,7 +37,7 @@ def test_execution_training_tracking(script_runner, experiments_path):
     input_group_name = TEST_EXPECTED_VOLUME_GROUPS[0]
     streamline_group_name = TEST_EXPECTED_STREAMLINE_GROUPS[0]
 
-    # Here, testing default values only. See dwi_ml.tests.test_trainer for more
+    # Here, testing default values only. See dwi_ml.unit_tests.test_trainer for more
     # various testing.
     logging.info("************ TESTING TRAINING ************")
     ret = script_runner.run('l2t_train_model.py',
