+++ scripts_python/tests/test_all_steps_ttst.py
@@ -7,10 +7,10 @@ import tempfile
 
 import torch
 
-from dwi_ml.tests.utils.expected_values import \
+from dwi_ml.unit_tests.utils.expected_values import \
     (TEST_EXPECTED_VOLUME_GROUPS, TEST_EXPECTED_STREAMLINE_GROUPS,
      TEST_EXPECTED_SUBJ_NAMES)
-from dwi_ml.tests.utils.data_and_models_for_tests import fetch_testing_data
+from dwi_ml.unit_tests.utils.data_and_models_for_tests import fetch_testing_data
 
 data_dir = fetch_testing_data()
 tmp_dir = tempfile.TemporaryDirectory()
@@ -41,7 +41,7 @@ def test_execution(script_runner, experiments_path):
     input_group_name = TEST_EXPECTED_VOLUME_GROUPS[0]
     streamline_group_name = TEST_EXPECTED_STREAMLINE_GROUPS[0]
 
-    # Here, testing default values only. See dwi_ml.tests.test_trainer for more
+    # Here, testing default values only. See dwi_ml.unit_tests.test_trainer for more
     # various testing.
     # Max length in current testing dataset is 108. Setting max length to 115
     # for faster testing. Also decreasing other default values.
