+++ scripts_python/tests/test_compute_loss_copy_previous.py
@@ -2,7 +2,7 @@
 # -*- coding: utf-8 -*-
 import os
 
-from dwi_ml.tests.utils.data_and_models_for_tests import fetch_testing_data
+from dwi_ml.unit_tests.utils.data_and_models_for_tests import fetch_testing_data
 
 data_dir = fetch_testing_data()
